Catch CommitErrors while rebasing
In Sentry we got reports of Gitlab::Git::CommitError occuring while
rebasing. This is nothing Gitaly can do anything about, so we catch this
as FailedPrecondition, the gRPC status code.

On the client side this error should be rescued from and turned into an
Gitlab::Git::CommitError again.

Closes gitlab-org/gitaly#1137

Internal links; not exhaustive list of sentry errors:
- https://sentry.gitlap.com/gitlab/gitaly-production/issues/158810/
- https://sentry.gitlap.com/gitlab/gitaly-production/issues/156359/
- https://sentry.gitlap.com/gitlab/gitaly-production/issues/156700/