Server implemenation ListNewCommits
Initially it was believed this RPC could be done by using other existing
RPCs. As tried in:
1. https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/20087
1. https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/6236

This wasn't as easy expected, and this RPC was required. The RPC wraps
the git rev-list command. To be precise, `git rev-list <rev> --not --all`
which analyses the revisions history to detect what commits do not occur
in any of the other branches, or HEAD. Given this is used after pushing,
before the branch is update, these are new commits.

This is a racy RPC, as git gc can be triggered right at this time. This is
for now not fixed. And its not this commits intention to fix it right now.

Vendoring of the proto is set to the branch of:
https://gitlab.com/gitlab-org/gitaly-proto/merge_requests/198. This will
be updated in later commits.

Part of https://gitlab.com/gitlab-org/gitaly/issues/1233