Fix file system type detection on *BSD and Darwin.
This commit adds support for DragonFly BSD, FreeBSD and OpenBSD to the
internal/helper/fstype package and fixes the implementation on Darwin.
On each of these platforms, fstype.FileSystem will call statfs(2) and
use the f_fstypename member of struct statfs to determine the file
system type. Unfortunately, the Go wrapper for this struct uses
differently named fields on OpenBSD and the other supported systems. For
this reason, a separate implementation for OpenBSD is included.