Enable feature flag toggling in node manager
To enable feature flag toggling in node manager, have a set of shards
that is not monitored and stays static. When an incoming request has the
feature flag, it will get directed to the version that does failover.
If it does not have the feature flag, it will get directed to the
primary as though the node manager did nothing.