Update publish script with release-tools tagging
Release-tools performs gitaly tagging, this MR updates the publish
script and deletes the release script.

* publish -> publish-gem

The script will check that the working dir is on the release tag, then
if there are no local changes it will build the gem and publish it to
rubygems.org

The script will now accept a RC version as parameter

* release

This script is no longer part of the process.