praefect: Convert replication latency metric to use a histogram vector
The replication latency metric is currently a simple histogram that
tracks the latency from starting a replication job until it's done. We
have multiple sets of replication jobs, though, which we can't discern
right now.

Convert the histogram to a histogram vector with a "type" label,
tracking what kind of replication job type we record.