praefect: Make replication event creation time available
While we already store a replication event's creation time in the
database, we do not make it available as part of the ReplJob structure.
As this is required to derive the delay until a job was started, this
commit makes the information available as a preparatory step for delay
metrics.