Make `lint` independent from golangci-lint-alpine image
Docker image that contains `golangci-lint` is not versioned
by Go version and does not make much sense for us by creating
report.
As result of moving to Go 1.14 we also need to update version
of `golangci-lint`. Otherwise it won't be run correctly.
Support for Go 1.14 included in
https://github.com/golangci/golangci-lint/releases/tag/v1.23.8
Changed to the latest 1.24.o release.