praefect: Implement new replication delay metric
As part of the replication, we have already implemented a latency metric
that measures how long a given replication job was running. This doesn't
account for the fact that there is an initial delay between queueing the
replication job and actually processing it and thus isn't sufficient to
deduce how long it took from queueing the job until it finished.

Add a new "delay" metric that fills this gap by recording how long it
took for a queued job to be started. Note that it doesn't record how
long it took in total from queuing until finalization, but this can be
trivially derived by adding up both "delay" and "latency" metrics.