transactions: Use properly seeded random number generator
By default, the `math/rand` package is seeded with a value of `1` and
thus produces the same sequence of random numbers on each start. This
isn't really what we want in the transactions manager, where we want to
get transaction identifiers which are as unique as possible. Especially
if restarting Praefect, we do not want to get the same sequence of
transaction identifiers but a new set to avoid the chance for collision.

Fix the issue by introducing a new `TransactionIdGenerator` interface
which allows users to supply their own generator. If none was chosen, a
default generator is used that uses `math.Rand` seeded with
`crypto/rand`. Note that the RNG returned by `math/rand.New()` is not
thread-safe, but as we already perform locking when generating
transaction identifiers we're guarded against concurrent access to the
RNG.