danger: Suggest merge request ID in the changelog
When creating a changelog entry, there's a slightly award dance where
the changelog entry needs to contain a merge request id which is unknown
at the time of generation. This create the situation where either a
developer guesses the next MR ID or has to create the MR, update the
changelog entry, and push again.

This change lets @GitalyBot suggest the MR ID, which can be applied by
the developer. Squashing the commit is left to the developer to do.