Refactoring of TLS usage
In order to implement proper support of TLS connections
for the praefect we require some refactoring of the existing
codebase.
Mainly it is about moving the things into other places and
fixing `config.Config` global var usage where is it accessible
as parameter.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/1698