danger: Update labels requiring no changelog
The "backstage" label is being deprecated in favor of a new set of
labels that are more specific. So let's update Danger to reflect that
and not complain about missing changelogs when any of the new labels is
set. At the same time, this commit also removes the old "backstage"
label so we stop using it.

Note that because of how GitLab interprets scoped labels, we need to now
quote all of our labels in the generated message.