metadata: Prefix transaction and server metadata with "gitaly"
Metadata and environment variables we inject and read for internal
purposes only are usually prefixed with the "gitaly-" for metadata and
"GITALY_" for environment variables. So let's do so for transactions and
Praefect server information too to not become inconsistent.

This also fixes some locations where we've used hardcoded strings to
test for metadata being present instead of using the constants exposed
by `internal/praefect/metadata`.