Remove dependency of remote package on config.Config
The remote package uses a global variable config.Config
that makes testing harder and error prone. It also reduces
flexibility of the configuration of the components.

As we are trying to break dependencies on the global vars
this change removes usage of the config.Config from the
remote package and substitutes it with passed in config.Cfg
parameter. As package is used by other parts of the service
it requires changes in a couple of other places to pass the
parameter from top to bottom.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699