command: Make passed in envs take higher priority
This commit makes passed in envs take a higher priority over envs
from environment, which avoids unexpected behaviors when we fall
into env conflicts for spawned commands. And we should always
respect args explicitly passed in anyway.

We are lucky that there are no such env conflicts for gitaly at
present, but we worth this additional safeguard.