add option for storing primary when creating a repository
Repository creation when using repository specific primaries does not
work as the creation goes through the usual repository mutator route
which fails due to the repository not having a primary. As a preliminary
step towards fixing that, this commit adds an option that can be used
to store a repository's primary node in the database when creating its
record.