praefect: Stop creating replication jobs on cleanup
The `RepositoryService.Cleanup` RPC performs a set of housekeeping tasks
for git repositories. Most importantly, it will remove stale lockfiles,
broken objects and broken references. One can thus consider this RPC to
be in the same class as `RepackFull` and `RepackIncremental` in that
they're doing essentially idempotent optimizations on a repository.
But while these RPCs get special handling in Praefect's coordinator,
`Cleanup` doesn't.

While the lack of special handling for this RPC doesn't impact
correctness, due to it being a mutator it will cause us to bump the
repository generation whenever it gets executed. This is essentially
useless as in non-corrupted repositories no user-visible change should
occur. But the biggest downside of it bumping the generation number is
that many mutating RPCs actually trigger a `Cleanup()` immediately
before or after the RPC. As a result, even if transactions are used, we
are forced to generate a replication job because of the `Cleanup()`,
rendering transactions a lot less useful.

Fix this issue by implementing special routing for those cleanup RPCs
similar to how repacks and garbage collection is handled, allowing us to
get rid of the generation bumping. As a result, we do not schedule
replication jobs anymore.

Note that originally, there had been two RPCs which got executed after
another mutator: `Cleanup()` and `ApplyGitattributes()`, which in
combination caused us to bump the generation twice. Turns out we were
implicitly relying on this generation bump for pushes though: when a
push was served, read requests inspecting the push must get routed to
the primary node. Otherwise e.g. the `/internal/allowed` checks are
broken because secondaries have a different quarantine directory for
received objects than the primary. We never hit this issue because of
the generation bump, which would've caused us to always route to the
primary. But with both RPCs being fixed via this commit and 77c0166c3
(repository: Use transactions when writing gitattributes, 2021-01-28),
all replicas are now considered up-to-date during the whole push, which
in combination with reads distribution causes failing access checks.

This bug has been fixed by implementing force-routing to primaries in
1102b0b67 (praefect: Implement force-routing to primary for node-manager
router, 2021-02-03) and 4d877d7d5 (praefect: Implement force-routing to
primary for per-repo router, 2021-02-03). But given that it also
requires the GitLab Rail supstream change edab619aa1f (gitaly: Fix
access checks with transactions and quarantine environments, 2021-02-05)
which has only been merged with v13.9, we have to wait until at least
v13.10 to land this fix.