return more detailed patch application errors from git2go.Apply
When applying a patch fails, we are currently returning fairly opaque
errors just saying merge conflict. Git's 'am' includes the number
of the patch and its subject in the error message as well. To maintain
error compatiblity with the Ruby and Go implementations of UserApplyPatch,
let's return the additional context with the errors so we can later
output matching error messages.