port UserApplyPatch to Go
This commit adds a Go port of UserApplyPatch.

UserApplyPatch is an RPC that parses a stream of patches in unix
mailbox format, applies them sequentially on top of the base commit
and sets the target branch to point to the resulting commit.

The original Ruby implementation basically just streamed the patches
to `git am`. The Go port tries to imitate the behavior as closely as
possible by supporting streaming, fallback three-way merging if
applying a patch fails and outputting the same error messages.

The RPC does not support setting a base branch. The Go implementation
follows the same 'default branch discovery' for determining the base
commit  as the Ruby implementation. Notably, neither support creating
the first branch in the repository.

The Go port is using libgit2 for the patch application. Git's and
libgit2's rename detection deviate slightly, with libgit2 not even
attempting to detect renames of small files. For this reason, an
existing test that triggers renaming behavior has been disabled.