extend test coverage of UserApplyPatch
UserApplyPatch' test coverage is very lacking at the moment. The
original tests are using gitlab/gitlab-test and applying various
patches to them. Some of the target branches have changed since then
with the patches not applying cleanly anymore.

This commit extends the suite with a bunch of new tests to nail down
the behavior in various cases. The test suite generates the state
needed on the fly to avoid testing against the moving target of
gitlab/gitlab-test.