export GetCommit test helper from git2go
We currently have a helper in git2go package which makes asserting
commits easier. It can't easily be moved to testhelpers as it uses
the git2go.Signature type. The best approach would likely be to
implement a GetCommit on localrepo.Repo and simply use require.Equal
to ensure the commmits match. However, the testhelper doesn't implement
robust enough parsing to promote it to a general method. We should
address this later by:

1. Implementing GetCommit on localrepo.Repo with complete parsing logic..
2. Move git2go.Signature to git.Signature as it is a general git
   structure, not only related to git2go.

To benefit from the helper in other tests, this commit exports the
helper from git2go until a better solution is put in place.