streamio: Return errors even if data was read
The implementation of `Read()` only returns read errors in case no data
was read at all. It's perfectly fine though for a reader to return both
an error and read data. Quoting the documentation of `io.Reader`:

    When Read encounters an error or end-of-file condition after
    successfully reading n > 0 bytes, it returns the number of bytes
    read. It may return the (non-nil) error from the same call or return
    the error (and n == 0) from a subsequent call.

So while we are allowed to defer the actual error to the next call of
`Read()`, we now rely on the underlying reader, which has to persist the
error and return it again on a subsequent call.

Fix the issue by just always returning both the number of bytes read and
the error.