fix TestRemovals flaking due to dependence on time
TestRemovals is testing removal of gitaly-ruby worker addresses from
Gitaly's gRPC resolver. There's a cooldown between removals which is by
the comment is supposed to provide back pressure for process restarts.

'add then remove but too fast' test case appears to be testing that the
removal cooldown is enforced. Due to the dependency on real time, this
can fail if the test is executed too flow and the second removal happens
after the 1 millisecond cooldown configured in the test.

This commit fixes the flaky test by removing the dependency on the actual
time and controlling the progression of time in the test. This way the test
works always regardless of the execution speed.

The test cases themselves are not smallest possible cases to demonstrate the
problem. I've left them as they were though and focused only on fixing the
time dependency and flakiness in this one.