streamio: Fix partial writes in `WriteTo()`
The `WriteTo()` implementation of the streamio `receiveReader` doesn't
correctly handle partial writes of its buffer: If only parts of the
current data have been written, it won't re-try to send remaining data.
This may potentially lead to gaps in the data that is currently getting
relayed and thus to hard-to-find bugs.

Fix this issue by looping over the buffer until no bytes remain for both
the left-over state as well as any new received state.