check the correct record in reconciler for assignments
Reconciler is not supposed to schedule jobs to unassigned nodes.
As it is, it doesn't bring outdated copies on unassigned storages
up to date but targets unassigned storages which don't have the
repository at all.

The bug is caused by checking for the assignments using a record that
might be null. We are checking the assignments using `storage_repositories`
row's virtual_storage and relative_path. The row in `storage_repositories`
is null if the node doesn't contain the repository. Finding assigments for
null virtual_storage and relative_path produces no rows which triggers the
fallback behavior of replicating everywhere as it looks like no assignments
exist.

This commit fixes the issue by checking against the `repositories` table's
row instead which is never null if a repository exists.