docs: Describe RemoveRepository behaviour
While reading and trying to understand how the repository was removed on
both the Rails and Gitaly side, I found it confusing that the RPC wasn't
describing behaviour that the caller wants to know.

This change documents the internals of an RPC so hopefully it's clearer
what Gitaly provides to callers.