delete repositories from unassigned storages
When a repository's replication factor is decreased, extra copies of
the repository are left on the now unassigned storages. This prevents
from reclaiming storage space on the cluster by reducing replication
factor of repositories and leaves unneeded cruft around. This commit
adds repository deletion logic in to Praefect's reconciler to reclaim
the storage space.

There are various checks in place to make sure we only delete the
unassigned copies of a repository when it is safe:

1. Assignments must be enabled for the repository by having one or
   more assignments stored in the database for the repository. This
   ensures cluster's which are not yet using variable replication
   factor remain unaffected.
2. All assigned storages must be on the same or later generation as
   the unassigned storage. This ensures we do not delete replicas on a
   later generation before the configured replication factor has been
   achieved on the assigned storages.
3. The repository must not be used as a target in any active replication
   job. This prevents duplicate deletion jobs from being scheduled while
   a deletion job is waiting in the queue for being processed. While no
   real changes should be targeting a replica that is eligible for deletion,
   we require the queue to be empty for the repository as additional safety
   measure.
3. The repository must not be used as a source in any active replication job.
   Reconciler may use an up to date unassigned replica as a source job for
   a replication. This ensures we don't break any already queued replications
   by deleting the repository. If that happened, we should still have another
   source node to recover the data from due to check number 2, but we require
   that no jobs use the repository as a source as an additional safety check.

Notably, we do not have knowledge of in-flight requests while we are scheduling
the jobs.

Write requests target only the assigned nodes, so in-flight write requests
should not be disturbed unless the node was unassigned during the in-flight
write. In transactional writes, this would equate to a failed vote. Primary
is always assigned, so this should not be a problem for non-transactional writes.
If the primary failed over and was unassigned, the write should fail anyway.

Reads are more problematic. Reads are distributed between all up to date storages.
As such, a read could be directed to an up to date unassigned node that is being
targeted for a deletion. The read could fail due to the deletion. This is likely
fine for now, given unassignments don't happen often. Once Praefect gains
rebalancing functionality, this will become more problematic as the rebalances
might happen more often. To improve the situation, we should do a follow-up where
we prioritize serving reads from assigned nodes to direct reads away from
unassigned nodes.

Similarly to update jobs, the storage must be healthy at the time of scheduling
to have deletion jobs scheduled. If the node is not healthy to process the jobs,
it doesn't make sense to schedule jobs for it.