operations: Return proper code if updating ref in UserCommitFiles races
Given that updating references is inherently races as any two processes
which update the same set of refs will fight over who is able to write
first, we need to treat errors when git-update-ref(1) fails specially.
We do in many cases by returning the FailedPrecondition error code,
which our monitoring infrastructure recognizes as an error which is fine
to be ignored.

The Go port of UserCommitFiles doesn't though: it will simply return an
unspecified error code. This has caused incidents because our monitoring
cannot handle this code and thus needs to treat it as a serious failure.

Fix this issue by returning the proper error code if updating refs
fails.