simplify test state generation in reconciler tests
Reconciler tests were using PostgresReplicationQueue to manipulate
replication job states. This does not work properly when one wants
to test historict states like cancelled, dead and completed as the
queue deletes them instead of storing them in the database. We may
have some historic state lying around though. To test against all
possible values, let's just simply set the database state as we want
it to be in the test with a custom query.