fix TestDial_Tracing/stream flaking
There was a previous attempt at synchronizing the test by waiting
for the server to close the stream. There is another race in the test
as well though.

The client first sends the stream close to the server and then goes off to finish
the client span. If the server receives the client stream closing before
the client managed to finish its span, the server will race finishing
its 'nested span' with the client's '/mwitkow.testproto/TestService/PingStream'
span.

This commit fixes the race by synchronizing the server's stream handler to only
proceed after the client has returned from the CloseSend method.