Commit creation with GitSetup
CreateCommit method is a replacement for the
CreateCommit function defined in the same file.
As part of this addition a couple of the new
configuration options were added as well.
The change also includes definition of the global
default value such as default commit message, etc.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699