featureflag: Remove UserFFBranch feature gate
Through b3a047a56 (Merge branch 'go_user_ff_branch_by_default' into 'master', 2021-01-27)
the UserFFBranch flag defaults to true. Which means that the feature
gate is up for removal if the changes are stable, which they are.

Thus this change removes the feature gate, and makes sure the tests
succeed. In the next release the Ruby implementation can be removed.