Makefile: Do not redownload protoc.zip if it's up-to-date
The protoc target is currently always re-downloading the protoc archive
whenever the Makefile changes. This is inefficient: in most cases it
didn't change anyway.

Fix this by instead using a `protoc.version` file, similar to how we
handle rebuilds of git and libgit2. This allows us to only redownload
protoc in case the protoc version changed.