Breaking circular dependency for RunGitalyServer
We can benefit from RunGitalyServer helper function
in out services tests. We would have a more or less
complete server setup when we test each service separately.
Before that change we were not able to use RunGitalyServer
inside service sub-packages because of a circular dependency.
It was caused by usage of RegisterAll function that registers
all services on the gRPC server. To get rid of this problem
RunGitalyServer now accepts a callback function that takes
a list of dependencies required to initialize and register
all required services. To reduce amount of parameters needed
they were assembled into a single Dependencies struct with
'Get'ter methods for each dependency.