Dependency on gitaly-ssh for specific configuration
gitaly-ssh is a binary that needs to be compiled as
a separate executable. Execution of the ConfigureGitalySSH
compiles it on run. It is not efficient if it is called
for each test run. That is why ConfigureGitalySSHBin was
added. It is compiles binary only once for the package
and re-uses it for the test runs by putting compiled
binary to proper location.