Refactoring of the TestSuccessfulFetchInternalRemote
The tests uses a single gitaly instance where two are
required. This change setups up two gitaly instances
and configures them accordingly (with hooks and gitaly-ssh).
The change also removes dependency on the global config.Config
variable.
A helper function cloneRepoAtStorage is removed as not used anymore.
ListenGitalySSHCalls refactored to use existing configuration without
creating a new one: the gitaly-ssh binary renamed to the gitaly-ssh-actual
but remains in the same directory and script placed in the same
directory as well.