Makefile: Use separate target to initialize git sources
We currently have a single target to both clone/update the git sources
and build them. As we're about to introduce the ability to fetch git
sources via curl(1), this commit splits up the recipe into two: first
the recipe to retrieve sources, and then the recipe to build them.

No functional changes are expected from this commit, but it eases the
introduction of the new way to fetch git sources.